requirements_tests.txt
coveralls
voluptuous
flake8
mock
pylint
pytest
pytest-cov
requests_mock
tox
